Config override UI updates
