+++ src/components/ciPipeline/ciPipeline.css
@@ -215,7 +215,7 @@
 
 .advanced-option-container {
     height: 100%;
-    width: 75%;
+    width: 100%;
 }
 
 .title-padding {
