+++ src/css/base.scss
@@ -1893,6 +1893,10 @@ textarea {
     width: 200px;
 }
 
+.w-212 {
+    width: 212px;
+}
+
 .w-250 {
     width: 250px;
 }
@@ -1982,6 +1986,10 @@ textarea {
 
 .h-36 {
     height: 36px !important;
+}
+
+.h-40 {
+    height: 40px !important;
 }
 
 .h-42 {
